fix: reconciliate main residents home with local resident home
chore: resolves https://github.com/UnlockedLabs/UnlockEdv2/pull/1169#issuecomment-4743152859
